catalogstudio-backend/requirements.txt
django>=5.0
djangorestframework
mysqlclient
django-cors-headers
python-dotenv
Pillow
yagmail
